Move libnfc-nci.conf to tone shared
All tone devices share the PN547C2 chip and the NCI config can be
unified.

This change also moves the files from /system/etc to /vendor/etc for
treble compliance. Since "nfc" is a system service, this means the
libnfc-nci.conf file in /vendor needs to be labeled in sepolicy and the
"nfc" domain needs to be granted access.

keyaki did not previously force tag polling, this commit also brings it
in line with the other tone devices by integrating this commit from
kagura/dora:

> Force tag polling for all available technologies
>
> Bring back the POLLING_TECH_MASK config entry
> and set it to force tag polling for all available techs.
>
> Without this config the chip will not work properly.
>
> Signed-off-by: Humberto Borba <humberos@omnirom.org>
> Change-Id: Iffc44208a0b5d26a879bd3b5d80df20181dde736